+++ tests/admin/app.py
@@ -11,11 +11,11 @@ class AdminTest(test.AppTestCase):
         self.assertIsInstance(admin, Admin)
         sitemap = admin.sitemap(app)
         self.assertTrue(sitemap)
-        self.assertEqual(len(sitemap), 2)
+        self.assertEqual(len(sitemap), 1)
         items = {}
         for site in sitemap:
             items.update(((item['title'], item) for item in site['items']))
-        self.assertEqual(len(items), 3)
+        self.assertEqual(len(items), 1)
         blog = items['Blog']
         self.assertEqual(blog['icon'], 'fa fa-book')
         self.assertEqual(blog['title'], 'Blog')
@@ -38,10 +38,5 @@ class AdminTest(test.AppTestCase):
 
     def test_edit_view(self):
         request = yield from self.client.get('/admin/blogs/1')
-        response = request.response
-        self.assertEqual(response.status_code, 200)
-
-    def test_edit_template_view(self):
-        request = yield from self.client.get('/admin/html_templates/1')
         response = request.response
         self.assertEqual(response.status_code, 200)
